st7735fb: Make FB native endian on little endian platforms
This is a quick unoptimized implementation of a shadow swap buffer
which is engaged on little endian platforms. It's not DMA safe
at this point nor is it particularly optimized during updates.
This is ok atm since the point is to be compatible with all the
user space code out there that assumes that all FBdevs are native
endian.

Now things like "cat foo.rgb565 > /dev/fb0" work as well as standard
apps like fbv with no byte swapping mods.

Signed-off-by: Matt Porter <mporter@ti.com>
Signed-off-by: Michel Pollet <buserror@gmail.com>