mxs: Allows specifying a bus number
SPI busses are by default enumerated down from 32k
and if spidev is used, it's impossible to predict
for sure the large number that will be allocated
and used in the /dev filename.

This patch introduces a "bus-num" property for
the .dts that let the device tree decide of
the device id, solving that problem.

Signed-off-by: Michel Pollet <buserror@gmail.com>